Reestructuración - New Dataset

+++ src/preprocessing/frame_processor.py
@@ -1,10 +1,11 @@
 import cv2
+import numpy as np
 
 def resize_frame(frame, target_size=(224, 224)):
     return cv2.resize(frame, target_size)
 
 def normalize_frame(frame):
-    return frame / 255.0
+    return frame.astype('float32') / 255.0
 
 def preprocess_frame(frame, target_size=(224, 224)):
     resized = resize_frame(frame, target_size)
